Changes:
Changed version code to 0.0.1 (Alpha)
Moved creation and writing of files to and asyncTask "EatBattery". This class will the activation of sensors and its use. Also writing the results to a file.
The app iis forced to always work work on portrait mode (i.e. hen rotating the phone nothigns happens) due to the fact that somme runtime changes can kill Asynctasks

Signed-off-by: Javier López Lumbreras <loplumja@hotmail.com>